transpile index better

+++ crates/comptime/husky-rust-transpilation/expect-files/examples/mnist-classifier/module_defn_rust_transpilation/digits/eight.rs
@@ -10,8 +10,8 @@ pub fn is_eight() -> OneVsAll {
     require!(let none = is_zero());
     require!(let none = is_seven());
     let upper_excess = major_connected_component().upper_mass() - major_connected_component().lower_mass();
-    if let none = major_connected_component().eff_holes().matches[1] {
-        if let none = major_connected_component().eff_holes().matches[0] {
+    if let none = major_connected_component().eff_holes().matches[1 as usize] {
+        if let none = major_connected_component().eff_holes().matches[0 as usize] {
             require!(false);
         }
         require!(false);
