+++ examples/mnist-classifier/target-rs/mnist-classifier/src/fermi.rs
@@ -9,7 +9,7 @@ pub fn fermi_match(concave_components: Leash<Vec<ConcaveComponent>>, templates: 
     let mut others = concave_components.collect_leashes();
     let mut matches: Vec<Option<Leash<ConcaveComponent>>> = vec![];
     for i in 0..templates.ilen() {
-        let template = templates[i];
+        let template = templates[i as usize];
         matches.push(others.pop_with_largest_opt_f32(template))
     }
     return FermiMatchResult::__constructor(matches, others);
@@ -19,7 +19,7 @@ impl FermiMatchResult {
     pub fn norm(self) -> f32 {
         let mut norm: f32 = 0.0f32;
         for i in 0..self.others.ilen() {
-            norm = norm.max(self.others[i].norm())
+            norm = norm.max(self.others[i as usize].norm())
         }
         return norm;
     }
@@ -27,7 +27,7 @@ impl FermiMatchResult {
     pub fn rel_norm(self) -> f32 {
         let mut norm: f32 = 0.0f32;
         for i in 0..self.others.ilen() {
-            norm = norm.max(self.others[i].rel_norm())
+            norm = norm.max(self.others[i as usize].rel_norm())
         }
         return norm;
     }
@@ -35,7 +35,7 @@ impl FermiMatchResult {
     pub fn angle_change_norm(self) -> f32 {
         let mut norm: f32 = 0.0f32;
         for i in 0..self.others.ilen() {
-            norm = norm.max(self.others[i].angle_change().abs())
+            norm = norm.max(self.others[i as usize].angle_change().abs())
         }
         return norm;
     }
