+++ examples/mnist-classifier/target-rs/mnist-classifier/src/line_segment_sketch/concave_component.rs
@@ -42,7 +42,7 @@ impl ConcaveComponent {
         let curve_ls = self.line_segment();
         let dp_norm = curve_ls.displacement().norm();
         for i in self.strokes.start()..self.strokes.end() {
-            let point = self.strokes[i].end;
+            let point = self.strokes[i as usize].end;
             let point_dist = curve_ls.dist_to_point(&point);
             if point_dist > hausdorff_norm {
                 hausdorff_norm = point_dist
@@ -53,9 +53,9 @@ impl ConcaveComponent {
 
     pub fn angle_change(self) -> f32 {
         let mut angle_change = 0.0f32;
-        let mut dp0 = self.strokes[self.strokes.start()].displacement();
+        let mut dp0 = self.strokes[self.strokes.start() as usize].displacement();
         for i in (self.strokes.start() + 1)..self.strokes.end() {
-            let dp = self.strokes[i].displacement();
+            let dp = self.strokes[i as usize].displacement();
             angle_change += dp0.angle_to(&dp, true);
             dp0 = dp
         }
@@ -69,7 +69,7 @@ impl ConcaveComponent {
         let mut ymin = start_point.y;
         let mut ymax = start_point.y;
         for i in self.strokes.start()..self.strokes.end() {
-            let point = self.strokes[i].end;
+            let point = self.strokes[i as usize].end;
             xmin = xmin.min(point.x);
             xmax = xmax.max(point.x);
             ymin = ymin.min(point.y);
